requirements.txt
fastapi
uvicorn[standard]
python-docx
python-multipart
jinja2
aiofiles
openai
